Environment split and wandb mode added, to reduce logging abundance
