+++ multi_agent_reinforcement_learning/algos/actor_critic_gnn.py
@@ -12,7 +12,7 @@ import torch.nn.functional as F
 from torch import nn
 from torch.distributions import Dirichlet
 
-from multi_agent_reinforcement_learning.envs.amod_env import AMoD
+from multi_agent_reinforcement_learning.envs.amod import AMoD
 from multi_agent_reinforcement_learning.algos.gnn_actor import GNNActor
 from multi_agent_reinforcement_learning.algos.gnn_critic import GNNCritic
 from multi_agent_reinforcement_learning.algos.gnn_parser import GNNParser
